+++ src/components/editor-page/table-of-contents/toc-slugify.ts
@@ -1,0 +1,11 @@
+/*
+ * SPDX-FileCopyrightText: 2021 The HedgeDoc developers (see AUTHORS file)
+ *
+ * SPDX-License-Identifier: AGPL-3.0-only
+ */
+
+export const tocSlugify = (content: string): string => {
+  return encodeURIComponent(content.trim()
+                                   .toLowerCase()
+                                   .replace(/\s+/g, '-'))
+}
